This code is not done.  I'm simply getting it ready for 10.6.6 and so far things are completely FUBAR on 10.5 with these changes.  For 10.5 the depricated interface runs good.

+++ mac/EventAction.h
@@ -24,10 +24,10 @@
 
 
 @interface EventAction : NSObject {
-	int downKeys[MAX_KEY_SIZE];
-	int upKeys[MAX_KEY_SIZE];
-	int downKeysCount;
-	int upKeysCount;
+	NSInteger downKeys[MAX_KEY_SIZE];
+	NSInteger upKeys[MAX_KEY_SIZE];
+	NSInteger downKeysCount;
+	NSInteger upKeysCount;
 	CGPoint center;
 	ProcessSerialNumber eventWindow;
